fix(facets,classification,suggest): remove flattened duplicate packages
Five packages to zero, failing packages 29 -> 24. 672 insertions against
15404 deletions, 88 files removed - the ratio is the finding.

Java has one faceting package, org.apache.lucene.facet, with 92 classes
across facetset, range, rangeonrange, sortedset, taxonomy and
taxonomy/writercache. facets/ covers 87 of them and preserves that
decomposition; facet/ covered 27, of which 26 were duplicates, and
flattened the subpackages into its root. It had zero importers outside
its own subtree. Removed - but class by class, because for several of the
duplicated root classes facet/ held the faithful copy: its Facets had all
five Java methods and its LabelAndValue had Java's Label/Value/Count, so
that fidelity moved into facets/ rather than being deleted with it.

The other four packages were the same shape. classification had two
catch-all files declaring non-generic ClassificationResult and Classifier
where Java has record ClassificationResult<T> and interface Classifier<T>,
plus "Simplified model for the purpose of the port" stubs for the naive
Bayes and kNN classifiers, plus a DocumentClassifier belonging to the
classification.document subpackage that Gocene already has, generically
and correctly. Removing those two files took 134 errors to 13.

suggest already had the faithful subpackage tree analyzing/document/fst/
spell/tst - suggest/document alone has all 34 Java classes - and the root
package re-declared it in "Simplified lookup logic" form. tests/suggest
was byte-identical to suggest apart from the package clause, and Lucene
has no org.apache.lucene.tests suggest classes at all.
AnalyzingInfixSuggester and BlendedInfixSuggester moved into
suggest/analyzing, which is org.apache.lucene.search.suggest.analyzing.

facets' single FacetResult redeclaration accounted for ~35 of its 95
errors. Java's is a final class with one constructor (dim, path, value,
labelValues, childCount); the second copy had a 1-arg constructor with
AddLabelValue mutators and a FacetResults collection Java has no
counterpart for. Which slice shape to keep was measured, not guessed:
[]*LabelAndValue appears in 11 sites across 6 files including green
subpackages, []LabelAndValue in 2 - both inside the outlier.

search/drill.go is removed: DrillDownQuery, DrillSidewaysQuery and
DrillSidewaysScorer are org.apache.lucene.facet classes misplaced into
search, and nothing anywhere referenced them.

Two genuinely missing Lucene methods are ported additively:
IndexWriter.getDocStats(), which DirectoryTaxonomyWriter needs, and
DocValues.getSortedSet(LeafReader, String).

A verification note worth keeping: the first orphan scan was file-name
based and produced false positives - FacetsResult and DrillSidewaysResult
look fabricated but are nested Java classes
(FacetsCollectorManager.FacetsResult, DrillSideways.DrillSidewaysResult).
Nothing was deleted until each name was re-checked against
(class|record|interface|enum) across the whole Java tree.

+++ sandbox/facet/utils/base_facet_builder.go
@@ -143,7 +143,7 @@ func (c *BaseFacetBuilderConfig) GetResult(
 	fr := &facets.FacetResult{
 		Dim:        c.Dimension,
 		Path:       c.Path,
-		Value:      overallValue,
+		Value:      float64(overallValue),
 		ChildCount: len(ords),
 	}
 	for _, ord := range ords {
